Set MAXPAGESIZE to 64k instead of 2MB on x86-64 NaCl
This brings x86-64 NaCl into line with x86-32 NaCl.

This reduces wastage of address space in dynamically-linked libraries
and executables.

It will also fix "strip" so that it can reduce x86-64 objects to
roughly the same size as x86-32 objects.  (This will reduce the size
of the IRT library shipped in Chromium.)

MAXPAGESIZE was set to 2MB on normal x86-64 ELF systems just in case
the ELF objects are used on a system that *only* supports large pages;
the resulting address space wastage is deemed acceptable because
normal x86-64 systems have address space to burn.

However, on NaCl, supporting 64k pages is part of our ABI, and since
the sandbox limits address space to 4GB, we do not have address space
to burn.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1073
TEST=
 1) check sizes of stripped IRT binaries
 2) check PT_LOAD addresses from "readelf -a" on libpthread.so

Review URL: http://codereview.chromium.org/6995023