Put rodata at a high address by default
We're going to move to the presumption that an IRT will be used.
So all links should put rodata above a gap to allow for IRT loading.

For the benefit of potential special cases, the new default behavior is
conditionalized so that the old default prevails if the symbol
__nacl_rodata_after_text is present in the link.

BUG= none
TEST= did local scons build and test run with this toolchain

R=eaeltsin@google.com

Review URL: http://codereview.chromium.org/7518013