Backport of upstream --enable-deterministic-archives, ar/ranlib -D/-U switches
This is a backport of my recent changes on the upstream binutils trunk.
It adds the -D switch to ranlib, same as ar's D modifier.
Then it adds -U to ar and ranlib, the converse of -D.
Then it adds --enable-deterministic-archives support to configure,
making the -D behavior the default.

This is all in service of getting bit-identical reproducible builds.
These are currently hamstrung by timestamps in .a files.

BUG= none
TEST= toolchain trybot

R=mseaborn@chromium.org,eaeltsin@google.com,khim@google.com,pasko@google.com

Review URL: http://codereview.chromium.org/8999020