Support --without-zlib configuration so binaries don't require libz.so
This is a backport of upstream changes (some already in upstream, some
pending upstream review) to fully support the --without-zlib configure
switch when building binutils.  Using this prevents compiling in the
supported for compressed DWARF sections and so avoids producing binaries
that have a runtime dependency on a libz.so.1 being installed on the system.

BUG= none
TEST= toolchain trybots

R=bradnelson@google.com

Review URL: http://codereview.chromium.org/9149014