Predefine 'naclret' macro in assembler
This is a hacky way to get the effect of:
	.macro naclret
	pop %ecx
	nacljmp %ecx
	.endm
without reading it from any input file.

This will let us change assembly code to use 'naclret' instead of
the pair of a pop and nacljmp.  In the next-generation x86 gcc,
this is defined as a normal macro and does the round-up version.